pyproject.toml
[tool.poetry]
name = "qa-guru-allure-testops-pytest"
version = "0.1.0"
description = "Intergrations Allure TestOps with Jenkins and Pytest"
authors = ["Rustam Dzhafarov <rustamd74@gmail.com>"]
readme = "README.md"
packages = [{include = "qa_guru_allure_testops_pytest"}]

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^7.1.3"
allure-pytest = "^2.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
markers = "api web  regress  smoke  critical"
